feat: add logic for history of style statistics

+++ models/history.go
@@ -7,7 +7,7 @@ import (
 type History struct {
 	gorm.Model
 	StyleID       uint
-	DailyViews    int
-	DailyInstalls int
-	DailyUpdates  int
+	DailyViews    int64
+	DailyInstalls int64
+	DailyUpdates  int64
 }
